Move canoegen2 to canoe_gen2
