+++ keyboards/percent/canoe_gen2/canoe_gen2.c
@@ -15,7 +15,7 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "canoegen2.h"
+#include "canoe_gen2.h"
 
 void keyboard_pre_init_kb(void) {
   setPinOutput(E6);
