+++ keyboards/percent/canoe_gen2/config.h
