+++ keyboards/percent/canoe_gen2/info.json
