+++ keyboards/percent/canoe_gen2/keymaps/default/keymap.c
