+++ keyboards/percent/canoe_gen2/keymaps/via/keymap.c
