+++ keyboards/percent/canoe_gen2/readme.md
@@ -8,10 +8,10 @@ Canoe Gen2 is a 65% from Percent with per key RGB and an adjustable foot angle. 
 
 Make example for this keyboard (after setting up your build environment):
 
-    make percent/canoegen2:default
+    make percent/canoe_gen2:default
 
 Flashing example for this keyboard:
 
-    make percent/canoegen2:default:flash
+    make percent/canoe_gen2:default:flash
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
