+++ keyboards/percent/canoe_gen2/rgb_matrix_kb.inc
