+++ keyboards/percent/canoe_gen2/rules.mk
